requirements.txt
Django==4.1
django-heroku>=0.3.1
dj-database-url
gunicorn>=20.1.0
pytz>=2022.4
requests>=2.25.1
whitenoise>=6.2.0
PyYAML>=6.0.0
cryptography>=38.0.1
psycopg2~=2.9.3
pandas
numpy
djangorestframework
